ADD: Observation report ADD: Process removals by age observation class that simulates data.

+++ Source/AgeingErrors/Children/Normal.h
@@ -37,7 +37,7 @@ protected:
 
 private:
   // members
-  Double                      cv_ = 0.0;
+  float                      cv_ = 0.0;
   unsigned                    k_;
 };
