+++ Source/AgeingErrors/Children/OffByOne.h
@@ -36,8 +36,8 @@ protected:
 
 private:
   // members
-  Double                      p1_ = 0.0;
-  Double                      p2_ = 0.0;
+  float                      p1_ = 0.0;
+  float                      p2_ = 0.0;
   unsigned                    k_;
 };
